+++ persistent-postgresql/test/PgInit.hs
@@ -1,4 +1,4 @@
-{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE OverloadedStrings, ScopedTypeVariables #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 
 module PgInit (
@@ -57,6 +57,7 @@ import Test.QuickCheck
 import Control.Monad (unless, (>=>))
 import Control.Monad.IO.Class
 import Control.Monad.IO.Unlift (MonadUnliftIO)
+import UnliftIO.Exception
 import Control.Monad.Logger
 import Control.Monad.Trans.Resource (ResourceT, runResourceT)
 import qualified Data.ByteString as BS
@@ -113,17 +114,35 @@ runConnInternal connType f = do
 
   flip runLoggingT (\_ _ _ s -> printDebug s) $ do
     logInfoN (if travis then "Running in CI" else "CI not detected")
-    case connType of
-      RunConnBasic -> withPostgresqlPool connString poolSize $ runSqlPool f
-      RunConnConf -> do
-        let conf = PostgresConf
-              { pgConnStr = connString
-              , pgPoolStripes = 1
-              , pgPoolIdleTimeout = 60
-              , pgPoolSize = poolSize
-              }
-            hooks = defaultPostgresConfHooks
-        withPostgresqlPoolWithConf conf hooks (runSqlPool f)
+    let go =
+            case connType of
+                RunConnBasic ->
+                    withPostgresqlPool connString poolSize $ runSqlPool f
+                RunConnConf -> do
+                    let conf = PostgresConf
+                          { pgConnStr = connString
+                          , pgPoolStripes = 1
+                          , pgPoolIdleTimeout = 60
+                          , pgPoolSize = poolSize
+                          }
+                        hooks = defaultPostgresConfHooks
+                    withPostgresqlPoolWithConf conf hooks (runSqlPool f)
+    -- horrifying hack :( postgresql is having weird connection failures in
+    -- CI, for no reason that i can determine. see this PR for notes:
+                    -- https://github.com/yesodweb/persistent/pull/1197
+    eres <- try go
+    case eres of
+        Left (err :: SomeException) -> do
+            eres' <- try go
+            case eres' of
+                Left (err' :: SomeException) ->
+                    if show err == show err'
+                    then throwIO err
+                    else throwIO err'
+                Right a ->
+                    pure a
+        Right a ->
+            pure a
 
 runConnAssert :: SqlPersistT (LoggingT (ResourceT IO)) () -> Assertion
 runConnAssert actions = do